+++ Sprocket/Core/ModelManager.h
@@ -20,9 +20,6 @@ private:
 public:
     ModelManager();
 
-    static std::shared_ptr<Mesh> LoadModel(const std::string& path);
-        // Loads a model without caching it.
-
     std::shared_ptr<Mesh> GetModel(const std::string& path);
         // Returns the model in the specified path. If it has not beed loaded
         // it will be loaded and stored internally.
